test(add some tests to extend coverage for michelson encoder errors): coverage for michelson encoder
add some tests to extend coverage for michelson encoder errors

fix #1976
